Add Llama2; refactors to llm
- rename modules/chatgpt to modules/llm
- let Llama2 and chatgpt share configuration
- remove redundant code in examples
