+++ src/recwizard/modules/llm/__init__.py
@@ -2,6 +2,6 @@ from .modeling_chatgpt_gen import ChatgptGen
 from .modeling_chatgpt_rec import ChatgptRec
 from .tokenizer_chatgpt import ChatgptTokenizer
 from .configuration_chatgpt_gen import ChatgptGenConfig
-from .configuration_chatgpt_rec import ChatgptRecConfig
+from .configuration_llm_rec import LLMRecConfig
 from .configuration_llm import LLMConfig
 from .modeling_llama_gen import LlamaGen
