+++ src/recwizard/modules/llm/configuration_llm.py
@@ -12,7 +12,12 @@ class LLMConfig(BaseConfig):
         model_name(str): The specified GPT model's name. 
     """
 
-    def __init__(self, max_gen_len: int = 0, prompt: dict = None, **kwargs):
+    def __init__(self, model_name='meta-llama/Llama-2-7b-chat-hf',
+                  max_gen_len: int = 0, 
+                  answer_type = 'movie',
+                  answer_mask = '<movie>',
+                  prompt: dict = None, 
+                  **kwargs):
         """
         Initializes the instance of this configuration.
 
@@ -21,10 +26,10 @@ class LLMConfig(BaseConfig):
         """
 
         super().__init__(**kwargs)
-        self.model_name = 'meta-llama/Llama-2-7b-chat-hf'
+        self.model_name = model_name
         self.max_gen_len = max_gen_len
-        self.answer_type = 'movie'
-        self.answer_mask = '<movie>'
+        self.answer_type = answer_type
+        self.answer_mask = answer_mask
         if prompt is not None:
             self.prompt = prompt
         else:
@@ -46,25 +51,3 @@ class LLMConfig(BaseConfig):
                     ' especially when the user is likely to be leaving.'
                 )
             }
-            # self.prompt = {
-            #     'role': 'user',
-            #     'content': (  
-            #         # ' The only answer is {self.answer_name}, do not give other answer.'
-            #         # ' Your sentence should be friendly and kind.'
-            #         # ' Do not reply me, and directly answer the question.'
-            #         # ' Do not say you cannot answer the question. Directly answer it.'
-            #         # ' Here is a history of dialogue: "{}",'
-            #         # f' and your  {self.answer_mask}.'
-            #         f' Now please generate a template for a program to fill. Do not directly give the name of the {self.answer_type}.'
-            #         f' Respond with a template where each {self.answer_type} is replaced with a {self.answer_mask} mask.'
-            #         # f' I will carefully review the previous dialogue before I respond to the user.'
-            #         # f' Due to regulations I am NOT allowed to return any name of {self.answer_type} directly.'
-            #         f' You should use a {self.answer_mask} mask to represent the name of the {self.answer_type},'
-            #         # f' if there is any {self.answer_type} in the answer.'
-            #         f' For example, if you intend to say: "Inception is a good movie".'
-            #         f' You should instead say "{self.answer_mask} is a good movie".'
-            #         f' Your response may or may not refer to a {self.answer_type}.'
-            #         f' Never mention any concrete name of {self.answer_type} in your response!'
-            #         # f' But notice that I should not use the form "({self.answer_mask})"'
-            #     )
-            # }
