+++ src/recwizard/modules/llm/modeling_chatgpt_rec.py
@@ -1,32 +1,33 @@
+import json
 import os
 from typing import Union, List
+
 import openai
 
 from recwizard import BaseModule
 from recwizard.modules.monitor import monitor
-from .configuration_chatgpt_gen import ChatgptGenConfig
 from .tokenizer_chatgpt import ChatgptTokenizer
+from .configuration_llm_rec import LLMRecConfig
 import logging
 
 logger = logging.getLogger(__name__)
 
 
-class ChatgptGen(BaseModule):
+class ChatgptRec(BaseModule):
     """
-    The generator implemented based on OpanAI's GPT models.
+    The recommender implemented based on OpanAI's GPT models.
 
     """
 
-    config_class = ChatgptGenConfig
+    config_class = LLMRecConfig
     tokenizer_class = ChatgptTokenizer
 
-    def __init__(self, config: ChatgptGenConfig, prompt=None, model_name=None, debug=False,
-                 **kwargs):
+    def __init__(self, config: LLMRecConfig, prompt=None, model_name=None, debug=False, **kwargs):
         """
         Initializes the instance based on the config file.
 
         Args:
-            config (ChatgptGenConfig): The config file.
+            config (ChatgptRecConfig): The config file.
             prompt (str, optional): A prompt to override the prompt from config file.
             model_name (str, optional): The specified GPT model's name. 
         """
@@ -48,7 +49,7 @@ class ChatgptGen(BaseModule):
             model_name (str, optional): The specified GPT model's name.
 
         Returns:
-             the instance.
+            (ChatgptRec): the instance.
         """
         if config is None:
             config = cls.config_class.from_pretrained(pretrained_model_name_or_path)
@@ -68,42 +69,37 @@ class ChatgptGen(BaseModule):
         Get a tokenizer.
 
         Returns:
-            (ChatgptTokenizer): the tokenizer.
+            (ChatgptRecTokenizer): the tokenizer.
         """
         # return lambda x: {'context': x}
         return ChatgptTokenizer()
 
     @monitor
-    def response(self, raw_input, tokenizer, recs: List[str] = None, max_tokens=None, temperature=0.5,
-                 model_name=None,
+    def response(self, raw_input: str, tokenizer=None, topk=3, max_tokens=None, temperature=0.5, model_name=None,
                  return_dict=False,
                  **kwargs):
         """
         Generate a template to response the processed user's input.
 
         Args:
-            raw_input (str): The user's raw input.
-            tokenizer (BaseTokenizer, optional): A tokenizer to process the raw input.
-            recs (list, optional): The recommended movies.
+            raw_input (dict): A dict that contains the question and its related information.
+            tokenizer (BaseTokenizer, optional): A tokenizer to process the question.
+            topk (int): The number of answers.
             max_tokens (int): The maximum number of tokens used for ChatGPT API.
             temperature (float): The temperature value used for ChatGPT API.
             model_name (str, optional): The specified GPT model's name.
             return_dict (bool): Whether to return a dict or a list.
 
-        Returns:
-            str: The template to response the processed user's input.
-        """
 
+        Returns:
+            list: The answers.
+        """
+        if topk == 0:
+            return {'output': [], 'links': []} if return_dict else []
         messages = tokenizer(raw_input)['messages']
-        # Add prompt at end
-        prompt = self.prompt.copy()
-        if recs is not None:
-            formatted_movies = ", ".join([f'{i + 1}. "{movie}"' for i, movie in enumerate(recs)])
-            prompt['content'] = prompt['content'].format(formatted_movies)
-        messages.append(prompt)
+        messages.append(self.prompt)
         if self.debug:
             logger.info('\ninput:', messages)
-
         res = openai.ChatCompletion.create(
             model=model_name or self.model_name,
             max_tokens=max_tokens,
@@ -113,10 +109,31 @@ class ChatgptGen(BaseModule):
 
         if self.debug:
             logger.info('\napi result:', res)
-        res_out = 'System: {}'.format(res)
+
+        try:
+            answers = json.loads(res)
+            if len(answers) < topk:
+                raise ValueError()
+            output = [answer['name'] for answer in answers][:topk]
+            links = [answer['uri'] for answer in answers][:topk]
+        except:
+            messages[-1] = self.config.backup_prompt
+            res = openai.ChatCompletion.create(
+                model=self.model_name,
+                max_tokens=max_tokens,
+                messages=messages,
+                temperature=temperature
+            )['choices'][0]['message']['content']
+            output = res.split(',')[:topk]
+            if len(output) < topk:
+                output = output + [''] * (topk - len(output))
+            links = [''] * topk
+
         if return_dict:
             return {
-                'input': messages,
-                'output': res_out,
+                "input": messages,
+                "output": output,
+                "links": {name: link for name, link in zip(output, links)}
             }
-        return res_out
+        else:
+            return output
